+++ src/app/background_queue/PingProcessor.ts
